Ref: extracting mapping helpers.
